+++ src/index.tsx
@@ -1,4 +1,7 @@
 import Button from './Button/index.js'
 import Typography from './Typography/index.js'
+import Input from './Input/index.js'
+import ModalBasic from './ModalBasic/index.js'
+import ButtonGroup from './ButtonGroup/index.js'
 
-export { Button, Typography }
+export { Button, Typography, ButtonGroup, Input, ModalBasic }